feature:监控模块任务数统计功能完成

+++ petrichor/starter/src/main/java/com/jalinyiel/petrichor/start/domain/TaskInfoSummary.java
@@ -11,12 +11,12 @@ public class TaskInfoSummary implements Serializable {
 
     private List<String> times;
 
-    private List<Integer> stringTaskCounts;
+    private List<Long> stringTaskCounts;
 
-    private List<Integer> listTaskCounts;
+    private List<Long> listTaskCounts;
 
-    private List<Integer> setTaskCount;
+    private List<Long> setTaskCount;
     
-    private List<Integer> mapTaskCount;
+    private List<Long> mapTaskCount;
 
 }
